backend/feat: surface CUMTA cumulative offer on FRFS search quotes
Add `offer` to FRFSQuoteAPIRes, populated for standalone (non-multimodal)
FRFS quotes in getFrfsSearchQuote via offerListCache + mkCumulativeOfferResp,
wrapped in withTryCatch so dynamic-logic failures degrade the optional offer
to Nothing instead of failing the quote. Multimodal journey legs are skipped
here (their offer is computed at the journey level). Split offer response
types into a SharedLogic.OfferTypes leaf to break the FRFSUtils ->
FRFSTicketService -> Offer -> JourneyModule import cycle; Offer re-exports
them. Add ClickhouseFlow to the FRFSSearchFlow alias.
