+++ Backend/app/rider-platform/rider-app/Main/rider-app.cabal
@@ -511,6 +511,7 @@ library
       SharedLogic.NyRegularSubscriptionHasher
       SharedLogic.NySubscription
       SharedLogic.Offer
+      SharedLogic.OfferTypes
       SharedLogic.OTP
       SharedLogic.PassExpiryReminder
       SharedLogic.PassRestore
