+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/FRFSTicketService.hs
@@ -90,6 +90,7 @@ import SharedLogic.FRFSStatus
 import SharedLogic.FRFSUtils
 import SharedLogic.FRFSUtils as FRFSUtils
 import qualified SharedLogic.IntegratedBPPConfig as SIBC
+import qualified SharedLogic.Offer as SOffer
 import qualified SharedLogic.PTCircuitBreaker as PTCircuitBreaker
 import qualified SharedLogic.Scheduler.Jobs.FRFSSeatHoldReaper as SeatHold
 import Storage.Beam.Payment ()
@@ -634,7 +635,7 @@ postFrfsSearchHandler (personId, merchantId) merchantOperatingCity integratedBPP
   return $ FRFSSearchAPIRes quotes searchReqId
 
 getFrfsSearchQuote :: (CallExternalBPP.FRFSSearchFlow m r, HasShortDurationRetryCfg r c) => (Kernel.Prelude.Maybe (Kernel.Types.Id.Id Domain.Types.Person.Person), Kernel.Types.Id.Id Domain.Types.Merchant.Merchant) -> Kernel.Types.Id.Id Domain.Types.FRFSSearch.FRFSSearch -> m [API.Types.UI.FRFSTicketService.FRFSQuoteAPIRes]
-getFrfsSearchQuote (mbPersonId, _) searchId_ = do
+getFrfsSearchQuote (mbPersonId, merchantId_) searchId_ = do
   personId <- fromMaybeM (InvalidRequest "Invalid person id") mbPersonId
   search <- QFRFSSearch.findById searchId_ >>= fromMaybeM (InvalidRequest "Invalid search id")
   integratedBppConfig <- SIBC.findIntegratedBPPConfigFromEntity search
@@ -676,6 +677,27 @@ getFrfsSearchQuote (mbPersonId, _) searchId_ = do
                in compare (maybe 0 (.amount) mbAdultPrice1) (maybe 0 (.amount) mbAdultPrice2)
           )
           routeFilteredQuotesWithCategories
+  let mbReprAdultPrice = listToMaybe sortedQuotesWithCategories >>= \(_, qcs) -> find (\c -> c.category == ADULT) qcs <&> (.price)
+  -- Cumulative offer is surfaced on standalone FRFS quotes only. When this search is
+  -- a leg of a multimodal journey, the offer is computed once at the journey level
+  -- (generateJourneyInfoResponse), so skip it here to avoid a redundant fetch.
+  -- This is keyed off the journey leg, NOT the config's platform type: a standalone
+  -- purchase can resolve a MULTIMODAL-platform config (e.g. Chennai bus has no
+  -- APPLICATION config), and the offer must still surface there.
+  mbOffer <-
+    if isJust mbJourneyLeg
+      then pure Nothing
+      else case mbReprAdultPrice of
+        Nothing -> pure Nothing
+        Just reprPrice ->
+          withTryCatch
+            "getFrfsSearchQuote:cumulativeOffer"
+            ( SOffer.offerListCache merchantId_ personId search.merchantOperatingCityId (FRFSUtils.getPaymentType False search.vehicleType) reprPrice Nothing
+                >>= \offersResp -> SOffer.mkCumulativeOfferResp search.merchantOperatingCityId offersResp [] Nothing
+            )
+            >>= \case
+              Left _ -> pure Nothing
+              Right mbResp -> pure mbResp
   mapM
     ( \(quote, quoteCategories) -> do
         let decodedRouteStations :: Maybe [FRFSRouteStationsAPI] = decodeFromText =<< quote.routeStationsJson
@@ -706,6 +728,7 @@ getFrfsSearchQuote (mbPersonId, _) searchId_ = do
               integratedBppConfigId = quote.integratedBppConfigId,
               stations = fromMaybe [] stations,
               observingFailures = Just observingFailures,
+              offer = mbOffer,
               ..
             }
     )
