+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/PartnerOrganizationFRFS.hs
@@ -589,6 +589,7 @@ mkQuoteRes (quote, quoteCategories) = do
         eventDiscountAmount = quote.eventDiscountAmount,
         integratedBppConfigId = quote.integratedBppConfigId,
         observingFailures = Nothing,
+        offer = Nothing,
         ..
       }
